+++ R/parameter_descriptions.R
@@ -1097,7 +1097,7 @@ NULL
 
 #' Parameter Description: Conditional Power Function (optimal conditional error design)
 #' @param conditionalPowerFunction
-#' A function accepting a single first-stage p-value and returning a numeric scalar strictly between 0 and 1. The function is evaluated separately for each p-value. This function should not be increasing in the first-stage p-value, otherwise monotonicity issues may occur.
+#' A function accepting a single first-stage p-value and returning a numeric scalar strictly between 0 and 1. The function is evaluated separately for each p-value. This function should not be increasing in the first-stage p-value, otherwise monotonicity issues may occur. With a conditional power function, the resulting conditional error function is not necessarily optimal for expected second-stage information.
 #' @name param_conditionalPowerFunctionOCEF
 #' @keywords internal
 NULL